+++ patreon_archiver/utils.py
@@ -131,8 +131,7 @@ def save_podcast(session: requests.Session, pdd: PostsData) -> SaveInfo:
         media_id = media_ref['id']
         with session.get(f'{MEDIA_URI}/{media_id}') as req:
             media: MediaData = req.json()['data']
-            download_url = media['attributes'].get('download_url')
-            if download_url:
+            if download_url := media['attributes'].get('download_url'):
                 file_name = Path(media['attributes'].get('file_name') or media_id).name
                 with session.get(download_url) as req2:
                     write_if_new(target_dir.joinpath(f'{media_id}-{file_name}'), req2.content, 'wb')
@@ -149,7 +148,9 @@ def save_podcast(session: requests.Session, pdd: PostsData) -> SaveInfo:
     return SaveInfo(post_data_dict=pdd, target_dir=target_dir)
 
 
-def process_posts(posts: Posts, session: requests.Session) -> Iterator[str | SaveInfo]:
+def process_posts(
+    posts: Posts, session: requests.Session, *, process_podcasts: bool = True
+) -> Iterator[str | SaveInfo]:
     """
     Process posts.
 
@@ -159,6 +160,9 @@ def process_posts(posts: Posts, session: requests.Session) -> Iterator[str | Sav
         The posts data from the API.
     session : requests.Session
         The requests session to use for downloads.
+    process_podcasts : bool
+        If ``True``, uses Patreon Archiver's handler to download podcasts and metadata. If
+        ``False``, the media URL is yielded.
 
     Yields
     ------
@@ -166,13 +170,10 @@ def process_posts(posts: Posts, session: requests.Session) -> Iterator[str | Sav
         If ``str`` it is a media URI. Otherwise it is a :py:class:`SaveInfo` object for an image or
         other post type.
     """
+    media_post_types = set(MEDIA_POST_TYPES) | ({'podcast'} if not process_podcasts else set())
     for post in posts['data']:
-        log.debug(
-            'Processing post: %s (%s)',
-            post['id'],
-            post['attributes']['post_type'],
-        )
-        if post['attributes']['post_type'] in MEDIA_POST_TYPES:
+        if post['attributes']['post_type'] in media_post_types:
+            log.debug('Sending URI: %s', post['attributes']['url'])
             yield post['attributes']['url']
         elif post['attributes']['post_type'] == 'image_file':
             yield save_images(session, post)
@@ -185,6 +186,8 @@ def process_posts(posts: Posts, session: requests.Session) -> Iterator[str | Sav
 def get_all_media_uris(
     campaign_id: str,
     session: requests.Session,
+    *,
+    process_podcasts: bool = True,
 ) -> Iterator[str]:
     """
     Get all media URIs for a given campaign ID.
@@ -195,6 +198,9 @@ def get_all_media_uris(
         The campaign ID to fetch posts for.
     session : requests.Session
         The requests session to use for API calls.
+    process_podcasts : bool
+        If ``True``, uses Patreon Archiver's handler for podcasts. If ``False``, yields podcast
+        media URLs for yt-dlp.
 
     Yields
     ------
@@ -204,14 +210,22 @@ def get_all_media_uris(
     r = session.get(POSTS_URI, params=get_shared_params(campaign_id))
     r.raise_for_status()
     posts: Posts = r.json()
-    yield from (x for x in process_posts(posts, session) if isinstance(x, str))
+    yield from (
+        x
+        for x in process_posts(posts, session, process_podcasts=process_podcasts)
+        if isinstance(x, str)
+    )
     next_uri = posts['links']['next']
     log.debug('Next URI: %s', next_uri)
     while next_uri:
         with session.get(next_uri) as req:
             req.raise_for_status()
             posts = req.json()
-            yield from (x for x in process_posts(posts, session) if isinstance(x, str))
+            yield from (
+                x
+                for x in process_posts(posts, session, process_podcasts=process_podcasts)
+                if isinstance(x, str)
+            )
             try:
                 next_uri = posts['links']['next']
                 log.debug('Next URI: %s', next_uri)